Update Subsilver2 style to 3.1
Imageset folder doesn't exist anymore.
The Hebrew buttons moved to a new language subfolder in
styles\subsilver2\theme .
The site_logo.gif image moved to styles\subsilver2\theme\images .

Consider updating on next stage:
* Update the text of the icon_contact_msnm.gif, from the old Hebrew
version "msnm", to the updated English version "WLN".
* Update the background color of site_logo.gif . In the updated English
version there is a gradient in the background, to match the header of
the style. In the Hebrew version the background is just a one color
background, without the gradient.